requirements.txt
Django>=4.0,<5.0
python-decouple>=3.6
djangorestframework>=3.12
cloudinary>=1.29
djangorestframework-simplejwt>=5.2
